requirements.txt
torch==2.3.1
torchvision==0.18.1
google-cloud-storage==2.18.2
tqdm==4.66.5
scikit-learn==1.5.2
langchain 
google-cloud-storage 
tqdm 
requests 
zipfile36
